+++ src/std/io_uring.rs
@@ -1,10 +1,4 @@
-use crate::{
-    error::{Error, PduError},
-    fmt,
-    std::unix::RawSocketDesc,
-    std::ParkSignal,
-    PduRx, PduTx,
-};
+use crate::{error::Error, fmt, std::unix::RawSocketDesc, std::ParkSignal, PduRx, PduTx};
 use core::{mem::MaybeUninit, task::Waker};
 use io_uring::{opcode, IoUring};
 use smallvec::{smallvec, SmallVec};
@@ -212,14 +206,6 @@ pub fn tx_rx_task_io_uring<'sto>(
                 loop {
                     match pdu_rx.receive_frame(&frame) {
                         Ok(_) => break,
-                        Err(Error::Pdu(PduError::NoWaker)) => {
-                            fmt::trace!(
-                                "No waker for received frame {:#04x}, retrying receive",
-                                frame_index
-                            );
-
-                            thread::yield_now();
-                        }
                         Err(e) => return Err(io::Error::other(e)),
                     }
                 }