+++ CHANGELOG.md
@@ -6,6 +6,16 @@ An EtherCAT master written in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+### Removed
+
+- **(breaking)** [#260](https://github.com/ethercrab-rs/ethercrab/pull/260) The `PduError::NoWaker`
+  variant has been removed as it is no longer used or returned by any EtherCrab method.
+
+### Fixed
+
+- [#260](https://github.com/ethercrab-rs/ethercrab/pull/260) Frames received before their future's
+  waker is registered will no longer trigger a `NoWaker` error.
+
 ## [0.5.1] - 2024-11-26
 
 ### Deprecated
@@ -439,8 +449,8 @@ An EtherCAT master written in Rust.
 - Initial release
 
 <!-- next-url -->
-[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-v0.5.1...HEAD
 
+[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-v0.5.1...HEAD
 [0.5.1]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-v0.5.0...ethercrab-v0.5.1
 [0.5.0]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-v0.4.2...ethercrab-v0.5.0
 [0.4.2]: https://github.com/ethercrab-rs/ethercrab/compare/ethercrab-v0.4.1...ethercrab-v0.4.2